Fix the minor issue with import items
