+++ packages/esm-admin-openconceptlab-app/src/previous-imports/import-overview/import-items.component.scss
@@ -1,0 +1,24 @@
+@import "~@openmrs/esm-styleguide/src/vars";
+@import "~carbon-components/src/globals/scss/vars";
+@import "~carbon-components/src/globals/scss/mixins";
+
+.pagination {
+  @include carbon--type-style('body-short-01');
+  background-color: $ui-02;
+  color: $text-02;
+  border: 1px solid $ui-03;
+  display: flex;
+}
+
+.tableBordered {
+  border: 1px solid $ui-03;
+}
+
+.tableDataRow > td {
+  padding-left: 1rem !important;
+  background-color: $ui-02 !important;
+}
+
+.tableHeaderRow > th {
+  background-color: $ui-03 !important;
+}
